async: convert methods to asyncio implementation when possible. Added "cached" decorator for class properties
When accessing/setting new properties, some of them require accessing to a file or decrypting a key. For performance purposes, the @cached decorator (for "get" methods) and it's pair @ucached (for "set" methods) store the correspondent new value inside a dictionary in which the key is the function name. When accessing the property, if there is a cached version available then it will be returned